feat(hwci): stronger tune search, confirmation, and session packaging
Split the auto-tuner monolith into a package and raise search/confirm
quality: soft multi-repeat DQ, climb on anchor-normalized scores,
optional high-throttle gate, min paired-delta confirmation with close-
call extra ABBA blocks, post-modes advance polish, adaptive anchors,
cool-only-when-hot, efficiency-argmax reporting, and pilot-card /
campaign export.
